requirements.txt
pandas
pyarrow
langchain
langchain-core
langchain-google-genai
google-generativeai
python-dotenv
sqlalchemy
pymongo
sqlite3
